+++ noodles-cram/src/crai/async.rs
@@ -9,7 +9,7 @@ use std::path::Path;
 
 use tokio::{fs::File, io};
 
-use super::Index;
+use super::{Index, Record};
 
 /// Reads the entire contents of a CRAM index.
 ///
@@ -34,4 +34,33 @@ where
 {
     let mut reader = File::open(src).await.map(Reader::new)?;
     reader.read_index().await
+}
+
+/// Writes a CRAM index to a file.
+///
+/// This is a convenience function and is equivalent to creating a file at the given path and
+/// writing the index.
+///
+/// # Examples
+///
+/// ```no_run
+/// # use std::io;
+/// #
+/// # #[tokio::main]
+/// # async fn main() -> io::Result<()> {
+/// use noodles_cram::crai;
+/// let index = crai::Index::default();
+/// crai::r#async::write("sample.cram.crai", &index).await?;
+/// # Ok(())
+/// # }
+/// ```
+
+pub async fn write<P>(dst: P, index: &[Record]) -> io::Result<()>
+where
+    P: AsRef<Path>,
+{
+    let mut writer = File::create(dst).await.map(Writer::new)?;
+    writer.write_index(index).await?;
+    writer.shutdown().await?;
+    Ok(())
 }